// packages/moim-web/app/actions/types.tsx
export enum EntityTypes {
  ADD_ENTITY = "ENTITY.ADD_ENTITY",
}
// ...
export enum ForumTypes {
  CHANGE_FORUM_ID = "FORUM.CHANGE_FORUM_ID",
  CHANGE_THREAD_ID = "FORUM.CHANGE_THREAD_ID",

  START_GET_THREAD_LIST = "FORUM.START_GET_THREAD_LIST",
  SUCCEED_GET_THREAD_LIST = "FORUM.SUCCEED_GET_THREAD_LIST",
  FAILED_GET_THREAD_LIST = "FORUM.FAILED_GET_THREAD_LIST",
  CLEAR_THREAD_LIST = "FORUM.CLEAR_THREAD_LIST",

  START_GET_COMMENT_LIST = "FORUM.START_GET_COMMENT_LIST",
  SUCCEED_GET_COMMENT_LIST = "FORUM.SUCCEED_GET_COMMENT_LIST",
  FAILED_GET_COMMENT_LIST = "FORUM.FAILED_GET_COMMENT_LIST",
  CLEAR_COMMENT_LIST = "FORUM.CLEAR_COMMENT_LIST",

  START_POST_COMMENT = "FORUM.START_POST_COMMENT",
  SUCCEED_POST_COMMENT = "FORUM.SUCCEED_POST_COMMENT",
  FAILED_POST_COMMENT = "FORUM.FAILED_POST_COMMENT",

  START_EDIT_COMMENT = "FORUM.START_EDIT_COMMENT",
  SUCCEED_EDIT_COMMENT = "FORUM.SUCCEED_EDIT_COMMENT",
  FAILED_EDIT_COMMENT = "FORUM.FAILED_POST_COMMENT",

  START_GET_THREAD = "FORUM.START_GET_THREAD",
  SUCCEEDED_GET_THREAD = "FORUM.SUCCEEDED_GET_THREAD",
  FAILED_GET_THREAD = "FORUM.FAILED_GET_THREAD",

  START_VOTE_THREAD = "FORUM.START_VOTE_THREAD",
  SUCCEEDED_VOTE_THREAD = "FORUM.SUCCEEDED_VOTE_THREAD",
  FAILED_VOTE_THREAD = "FORUM.FAILED_VOTE_THREAD",

  START_GET_THREAD_VOTES = "FORUM.START_GET_THREAD_VOTES",
  SUCCEEDED_GET_THREAD_VOTES = "FORUM.SUCCEEDED_GET_THREAD_VOTES",
  FAILED_GET_THREAD_VOTES = "FORUM.FAILED_GET_THREAD_VOTES",

  START_VOTE_REPLY = "FORUM.START_VOTE_REPLY",
  SUCCEEDED_VOTE_REPLY = "FORUM.SUCCEEDED_VOTE_REPLY",
  FAILED_VOTE_REPLY = "FORUM.FAILED_VOTE_REPLY",

  START_GET_REPLY_VOTES = "FORUM.START_GET_REPLY_VOTES",
  SUCCEEDED_GET_REPLY_VOTES = "FORUM.SUCCEEDED_GET_REPLY_VOTES",
  FAILED_GET_REPLY_VOTES = "FORUM.FAILED_GET_REPLY_VOTES",

  OPEN_VOTED_DIALOG = "FORUM.OPEN_VOTED_DIALOG",
  CLOSE_VOTED_DIALOG = "FORUM.CLOSE_VOTED_DIALOG",
  CLEAR_VOTED_DIALOG_DATA = "FORUM.CLEAR_VOTED_DIALOG_DATA ",

  ADD_THREAD = "FORUM.ADD_THREAD",
  CLEAR_CURRENT_THREAD = "FORUM.CLEAR_CURRENT_THREAD",

  START_DELETE_THREAD = "FORUM.START_DELETE_THREAD",
  SUCCEED_DELETE_THREAD = "FORUM.SUCCEED_DELETE_THREAD",
  FAILED_DELETE_THREAD = "FORUM.FAILED_DELETE_THREAD",

  START_DELETE_REPLY = "FORUM.START_DELETE_REPLY",
  SUCCEED_DELETE_REPLY = "FORUM.SUCCEED_DELETE_REPLY",
  FAILED_DELETE_REPLY = "FORUM.FAILED_DELETE_REPLY",

  CHANGE_FILTER_OPTION = "FORUM.CHANGE_FILTER_OPTION",

  OPEN_NEW_ITEM_SNACKBAR = "FORUM.OPEN_NEW_ITEM_SNACKBAR",
  CLOSE_NEW_ITEM_SNACKBAR = "FORUM.CLOSE_NEW_ITEM_SNACKBAR",
  SET_NEW_ITEM_SNACKBAR_DIRECTION = "FORUM.SET_NEW_ITEM_SNACKBAR_DIRECTION",

  CLEAR_COMMENT_EDIT_STATE = "FORUM.CLEAR_COMMENT_EDIT_STATE",
  CHANGE_COMMENT_EDIT_STATE = "FORUM.CHANGE_COMMENT_EDIT_STATE",

  START_PIN_POST = "CHANNEL.START_PIN_POST",
  SUCCEEDED_PIN_POST = "CHANNEL.SUCCEEDED_PIN_POST",
  FAILED_PIN_POST = "CHANNEL.FAILED_PIN_POST",

  START_GET_PINNED_POST_LIST = "CHANNEL.START_GET_PINNED_POST_LIST",
  SUCCEEDED_GET_PINNED_POST_LIST = "CHANNEL.SUCCEEDED_GET_PINNED_POST_LIST",
  FAILED_GET_PINNED_POST_LIST = "CHANNEL.FAILED_GET_PINNED_POST_LIST",

  START_DELETE_PINNED_POST = "CHANNEL.START_DELETE_PINNED_POST",
  SUCCEEDED_DELETE_PINNED_POST = "CHANNEL.SUCCEEDED_DELETE_PINNED_POST",
  FAILED_DELETE_PINNED_POST = "CHANNEL.FAILED_DELETE_PINNED_POST",

  START_ARRANGE_PINNED_POST_LIST = "CHANNEL.START_ARRANGE_PINNED_POST_LIST",
  SUCCEEDED_ARRANGE_PINNED_POST_LIST = "CHANNEL.SUCCEEDED_ARRANGE_PINNED_POST_LIST",
  FAILED_ARRANGE_PINNED_POST_LIST = "CHANNEL.FAILED_ARRANGE_PINNED_POST_LIST",

  OPEN_EDIT_PINNED_POST_LIST_DIALOG = "FORUM.OPEN_EDIT_PINNED_POST_LIST_DIALOG",
  CLOSE_EDIT_PINNED_POST_LIST_DIALOG = "FORUM.CLOSE_EDIT_PINNED_POST_LIST_DIALOG",

  UPDATE_HIGHLIGHT_THREAD_ID = "FORUM.UPDATE_HIGHLIGHT_THREAD_ID",
  CLEAR_HIGHLIGHT_THREAD_ID = "FORUM.CLEAR_HIGHLIGHT_THREAD_ID",

  START_SEARCH_THREADS = "FORUM.START_SEARCH_THREADS",
  SUCCEED_SEARCH_THREADS = "FORUM.SUCCEED_SEARCH_THREADS",
  FAILED_SEARCH_THREADS = "FORUM.FAILED_SEARCH_THREADS",
  CLEAR_SEARCH_THREADS = "FORUM.CLEAR_SEARCH_THREADS",

  START_FETCH_SCRAP = "FORUM.START_FETCH_SCRAP",
  SUCCEED_FETCH_SCRAP = "FORUM.SUCCEED_FETCH_SCRAP",
  FAILED_FETCH_SCRAP = "FORUM.FAILED_FETCH_SCRAP",
}
// ...
export enum ForumEditorTypes {
  CLEAR_POSTING_THREAD = "FORUM_EDITOR.CLEAR_POSTING_THREAD",
  START_POSTING_THREAD = "FORUM_EDITOR.START_POSTING_THREAD",
  SUCCEEDED_POSTING_THREAD = "FORUM_EDITOR.SUCCEEDED_POSTING_THREAD",
  FAILED_POSTING_THREAD = "FORUM_EDITOR.FAILED_POSTING_THREAD",

  START_UPDATE_THREAD = "FORUM_EDITOR.START_UPDATE_THREAD",
  SUCCEEDED_UPDATE_THREAD = "FORUM_EDITOR.SUCCEEDED_UPDATE_THREAD",
  FAILED_UPDATE_THREAD = "FORUM_EDITOR.FAILED_UPDATE_THREAD",
}